requirements.txt
certifi==2017.4.17
chardet==3.0.3
click==6.7
idna==2.5
pushbullet.py==0.10.0
python-magic==0.4.13
requests==2.17.3
six==1.10.0
urllib3==1.21.1
websocket-client==0.40.0
